# Major update to the grid picker to improve it's performance + other bug fixes