--- a/theme/fieldeditors.less
+++ b/theme/fieldeditors.less
@@ -68,23 +68,32 @@
 }
 
 .blocklyGridPickerPadder {
+    outline: none;
     box-shadow: 0px 0px 8px 1px rgba(0, 0, 0, .3)
 }
 
 .blocklyWidgetDiv .blocklyGridPickerMenu {
     display: table;
     outline: none;
+    border-spacing: 7px;
 }
 
 .blocklyGridPickerScroller {
     outline: none;
     padding: 4px;
-    border-radius: 10px;
+    border-radius: 4px;
+    position: relative;
+    -webkit-overflow-scrolling: touch;
 }
 
 .blocklyGridPickerPadder {
-    border-radius: 10px;
+    border-radius: 4px;
     outline: none;
+    position: relative;
+}
+
+.blocklyGridPickerPadder .ui.input i.search.icon {
+    margin-top: -0.2rem;
 }
 
 .blocklyWidgetDiv .blocklyGridPickerRow {
@@ -98,15 +107,23 @@
 
 .blocklyWidgetDiv .blocklyGridPickerMenu .goog-menuitem-highlight, .blocklyWidgetDiv .blocklyGridPickerMenu .goog-menuitem-hover {
     background: #d6e9f8;
+    box-shadow: 0px 0px 0px 4px rgba(255, 255, 255, 0.2);
 }
 
 .blocklyWidgetDiv .blocklyGridPickerMenu .goog-option {
-    border: solid 4px black;
-    border-radius: 10px;
+    border: solid 1px black;
+    border-radius: 4px;
+    color: #fff;
+    font-size: 12pt;
+    font-weight: bold;
     display: table-cell;
-    padding: 4px;
+    padding: 8px;
     text-align: center;
     vertical-align: top;
+    -webkit-user-select: none;
+    -moz-user-select: -moz-none;
+    -ms-user-select: none;
+        user-select: none;
 }
 
 .blocklyWidgetDiv .blocklyGridPickerMenu .floatLeft {
@@ -114,7 +131,6 @@
 }
 
 .blocklyWidgetDiv .blocklyGridPickerMenu .goog-option.goog-option-selected {
-    outline: solid 3px #fc3;
     position: relative;
 }
 
@@ -124,6 +140,18 @@
 
 .blocklyGridPickerTooltip {
     z-index: @gridPickerTooltipZIndex;
+}
+
+.blocklyGridPickerSelectedBar {
+    display: flex;
+    padding-top: 5px;
+    justify-content: space-between;
+}
+
+.blocklyGridPickerSelectedImage {
+    padding: 3px;
+    display: inline-block;
+    vertical-align: middle;
 }
 
 .ui.input input.blocklyGridPickerSearchBar {