@pelikhan 2018-10-11
Awesome improvements!

Get Outlook for iOS<https://aka.ms/o0ukef>

________________________________
From: Sam El-Husseini <notifications@github.com>
Sent: Wednesday, October 10, 2018 6:14 PM
To: Microsoft/pxt
Cc: Subscribed
Subject: [Microsoft/pxt] Major update to the grid picker to improve it's performance + other bug fixes (#4798)


Improve the performance of the grid picker: (current: ~1400ms to load)

  *   Add intersection observer support for images (shaved off 200ms)
  *   Remove closure dependency (~800ms)
  *   Add tooltips in a more efficient way (working with closure meant that we had to add tooltips as a post process) (~100ms)
It comes down to ~200ms to load

Also in this PR:

  *   Support a better touch experience by adding a selected bar. Selected bar only appears if there are tooltips, otherwise, we select on click.
  *   Updated look of grid picker with no images (micro:bit)
  *   Bug fixes with positioning of grid picker as well as bug fixes with hiding the tooltips after close.

Selected bar:
Only shows up on mobile devices and makes it so that the experience of selecting an item is in two steps. A selection which shows you the name of the image, and then a select / cancel button to select.
[screen shot 2018-10-10 at 6 08 26 pm]<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fuser-images.githubusercontent.com%2F16690124%2F46774514-a2f75600-ccb7-11e8-9833-ef3bc5e098ed.png&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763303023&sdata=O9Gc6yTttB8vkV25tbV8KTKCgCniSxubV7Bu275%2F2as%3D&reserved=0>

Updated look for grid pickers without an image:
[screen shot 2018-10-10 at 6 12 49 pm]<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fuser-images.githubusercontent.com%2F16690124%2F46774612-2e70e700-ccb8-11e8-81d9-9594ccce6034.png&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763303023&sdata=cRKZHp0UNqdRZ%2BhneVjddOOMfaq51KlqApfJWK5i%2Bms%3D&reserved=0>

Tested:

  *   micro:bit grid pickers for pins
  *   Minecraft grid pickers, ones with a set number of rows and ones without, same for ones with searchbar and without
  *   Devices: Mac, Windows, iPad
  *   Browsers: Chrome and Edge, Safari, Chrome Mobile

Pull in a pxt-blockly build with these two PRs:
Microsoft/pxt-blockly#125<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt-blockly%2Fpull%2F125&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763313031&sdata=XfeO8T1QSgTv5eoG0n2WkbRpT22LaClbvKPvZYvJKMQ%3D&reserved=0>
Microsoft/pxt-blockly#126<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt-blockly%2Fpull%2F126&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763323040&sdata=VIRoxrrCtmsOXEvD35tOLpYhiNjtfhEZLYO2aOcrnKk%3D&reserved=0>

________________________________
You can view, comment on, or merge this pull request online at:

  https://github.com/Microsoft/pxt/pull/4798<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763323040&sdata=NJDMYdmGXat41iQvCqlzafS7Qu%2FaD0RuOExvNXczda8%3D&reserved=0>

Commit Summary

  *   Major update to the grid picker to improve it's performance

File Changes

  *   M localtypings/blockly.d.ts<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798%2Ffiles%23diff-0&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763333049&sdata=Ky77kI6aev0UvzonuHRT4KAE6CD%2F%2BqX%2FAq2n3Qd%2BuWI%3D&reserved=0> (1)
  *   M pxtblocks/fields/field_gridpicker.ts<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798%2Ffiles%23diff-1&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763343057&sdata=cfxVAEo4WwjNSGynh%2F5aPexwgWtfabDC1lXxiBuycFE%3D&reserved=0> (738)
  *   M pxtblocks/fields/field_imagedropdown.ts<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798%2Ffiles%23diff-2&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763343057&sdata=KgaAeuqGmWGrJW9fWddj0MjB79mjAR03xqSVitatcuA%3D&reserved=0> (3)
  *   M theme/fieldeditors.less<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798%2Ffiles%23diff-3&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763353066&sdata=w8oSmyJZK52p%2F7DUu1Iac1fThMxfynWHQlMOcCD5Aek%3D&reserved=0> (40)
  *   M webapp/public/blockly/blockly_compressed.js<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798%2Ffiles%23diff-4&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763363075&sdata=%2B%2BX7VjScV3VumZbnTYTBQ6bL0JzhuQlCq6OHRAcu4t0%3D&reserved=0> (5)

Patch Links:

  *   https://github.com/Microsoft/pxt/pull/4798.patch<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798.patch&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763363075&sdata=paAM%2Bo68LUIqV07swvCNSBeI2ESHg5s%2BTE90lHt%2FYTs%3D&reserved=0>
  *   https://github.com/Microsoft/pxt/pull/4798.diff<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798.diff&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763373083&sdata=73%2FrOxWbvYiFDZGHNBIPKck%2BzltEv9uE6cvzbwld%2BsA%3D&reserved=0>

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F4798&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763383092&sdata=G9gVEnys%2BFd%2FTFEeVtIm55YaG2JoYrJWXQM%2B%2F36bt6w%3D&reserved=0>, or mute the thread<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FAD-4KWeeW52XLeT2aB9Ncl1PyP0gGVSUks5ujpt5gaJpZM4XWmSh&data=02%7C01%7Cjhalleux%40microsoft.com%7C7613e7ccc169490887cf08d62f16e7c2%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636748172763383092&sdata=LQR4uo4mcCgrr6c2oHlfnoXw53Z7Ijvdi2daWO8fCO8%3D&reserved=0>.